correct the LICENSE copyright to my name
as the Open Stick Community is not a legal entity, and I have not
transferred copyright to anyone/thing else, and because I am to date the
only author in this repository, I believe it is clearer to all and
stronger to just be explicit here and put my name in the license
document. this will be followed up by a SPDX update to the files
accordingly.

Signed-off-by: Brian S. Stephan <bss@incorporeal.org>